+++ CHANGELOG.md
@@ -9,8 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 ### Changed
-* Refactored ParallelVariableAnnealingLength to eliminate redundant code
-* Refactored constructors of window limited permutation mutation operators to eliminate redundant code
+* Refactored ParallelVariableAnnealingLength to eliminate redundant code.
+* Refactored constructors of window limited permutation mutation operators to eliminate redundant code.
+* Refactored BlockMoveIterator, WindowLimitedBlockMoveIterator, and BlockInterchangeIterator to remove unnecessary conditions.
 
 ### Deprecated